# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger LawyerUp - OpenAPI 3.0",
    "description": "This is a LawyerUp Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "paths": {
...
    "/user/me": {
      "get": {
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting information of the current user.",
        "description": "No input required for this api, just needed to be authenticated.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/user": {
      "patch": {
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting information of the current user.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "designation": {
                    "type": "string",
                    "example": "string"
                  },
                  "profilePicture": {
                    "type": "string",
                    "example": "string"
                  },
                  "roleAtCompany": {
                    "type": "string",
                    "example": "string"
                  },
                  "stepsCompleted": {
                    "type": "number",
                    "example": "0"
                  },
                  "firmName": {
                    "type": "string",
                    "example": "string"
                  },
                  "firmWebsite": {
                    "type": "string",
                    "example": "string"
                  },
                  "billingStructure": {
                    "type": "array",
                    "example": ["FIXED_FEE", "ON_CONTINGENCY", "PER_HOUR"]
                  },
                  "ratePerHourMin": {
                    "type": "number",
                    "example": "1"
                  },
                  "ratePerHourMax": {
                    "type": "number",
                    "example": "2"
                  },
                  "represent": {
                    "type": "array",
                    "example": ["INDIVIDUAL", "COMPANIES"]
                  },
                  "practiceAreas": {
                    "type": "object",
                    "example": ["string"]
                  },
                  "practicingLawSince": {
                    "type": "date",
                    "example": "dateString"
                  },
                  "licenses": {
                    "type": "array",
                    "example": [
                      {
                        "state": "stateId",
                        "licenseNumber": "Li-123"
                      }
                    ]
                  },
                  "locationPermitted": {
                    "type": "array",
                    "example": ["stateId"]
                  },
                  "biography": {
                    "type": "string",
                    "example": "string"
                  },
                  "languages": {
                    "type": "array",
                    "example": [
                      {
                        "language": "languageId",
                        "oralProficiency": "NATIVE",
                        "writtenProficiency": "NATIVE"
                      }
                    ]
                  },
                  "socialURLs": {
                    "type": "array",
                    "example": [
                      {
                        "kind": "string",
                        "url": "string"
                      }
                    ]
                  },
                  "educations": {
                    "type": "array",
                    "example": [
                      {
                        "university": "string",
                        "degree": "string",
                        "completedAt": "2023-04-05T22:10:43.275Z",
                        "certificates": ["string"]
                      }
                    ]
                  },
                  "publications": {
                    "type": "array",
                    "example": [
                      {
                        "title": "string",
                        "link": "string",
                        "imageUrl": "string"
                      }
                    ]
                  },
                  "awards": {
                    "type": "array",
                    "example": [
                      {
                        "title": "string",
                        "date": "2023-04-05T22:10:43.275Z"
                      }
                    ]
                  },
                  "associations": {
                    "type": "array",
                    "example": [
                      {
                        "title": "string",
                        "date": "2023-04-05T22:10:43.275Z"
                      }
                    ]
                  },
                  "experiences": {
                    "type": "array",
                    "example": [
                      {
                        "title": "title 1",
                        "description": "Description",
                        "practiceAreas": ["categoryId"],
                        "industry": "industry name",
                        "court": "court 1"
                      }
                    ]
                  },
                  "businesses": {
                    "type": "array",
                    "example": [
                      {
                        "name": "Business name",
                        "businessLogo": "string",
                        "isGeneralCounselor": true,
                        "businessType": "LLC",
                        "registeredDate": "2023-04-05T22:10:43.275Z",
                        "EIN": "EIN 1",
                        "businessPhone": "45 66539878",
                        "website": "www.example.com",
                        "businessDescription": "business description",
                        "addresses": [
                          {
                            "address": "address detail",
                            "officeType": "string",
                            "addressType": "REGISTERED",
                            "postalCode": "00112",
                            "city": "NY",
                            "state": "stateId",
                            "country": "country"
                          }
                        ]
                      }
                    ]
                  },
                  "addresses": {
                    "type": "array",
                    "example": [
                      {
                        "address": "address detail",
                        "officeType": "string",
                        "addressType": "REGISTERED",
                        "postalCode": "00112",
                        "city": "NY",
                        "state": "stateId",
                        "country": "country"
                      }
                    ]
                  },
                  "subscription": {
                    "type": "string",
                    "example": "STANDARD"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting user by userId.",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Unique id of the user which you want to retrieve.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/user/my-searches": {
      "post": {
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting searches of the current user.",
        "description": "No input required for this api, just needed to be authenticated.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "type": "number"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "filters": {
                    "type": "object",
                    "example": {
                      "isBookMarkedOnly": true
                    }
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}